Fixed the test for packages

+++ oqe_install.py
@@ -16,8 +16,20 @@ OQ_CFG = os.path.join(VENV, 'openquake.cfg')
 OQ = '/usr/bin/oq'
 DBSERVER_PORT = 1907
 
+PACKAGES = '''It looks like you have an installation from packages.
+Please remove it with `sudo apt remove python3-oq-engine`
+on Debian derivatives or with `sudo yum remove python3-oq-engine`
+on Red Hat derivatives. If it does not work, just remove everything with
+sudo rm -rf /opt/openquake /etc/openquake/openquake.cfg /usr/bin/oq
+'''
+
 
 def before_checks():
+    # check python version
+    if sys.version_info[:2] < (3, 6):
+        sys.exit('Error: you need at least Python 3.6, but you have %s' %
+                 '.'.join(map(str, sys.version_info)))
+
     # check if there is a DbServer running
     sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
     try:
@@ -27,14 +39,16 @@ def before_checks():
     if errcode == 0:  # no error, the DbServer is up
         sys.exit('There is DbServer running on port %d from a previous '
                  'installation. Please run `oq dbserver stop`' % DBSERVER_PORT)
+
+    # check if there is an installation from packages
+    if os.path.exists('/etc/openquake/openquake.cfg'):
+        sys.exit(PACKAGES)
+
     user = getpass.getuser()
     if user != 'root':
         sys.exit('Error: you cannot run this script unless you are root. If '
                  'you do not have root permissions, you can always install '
                  'the engine in single-user mode.')
-    if sys.version_info[:2] < (3, 6):
-        sys.exit('Error: you need at least Python 3.6, but you have %s' %
-                 '.'.join(map(str, sys.version_info)))
     if os.path.exists(OQ) and os.readlink(OQ) != '%s/bin/oq' % VENV:
         sys.exit('Error: there is already a link %s->%s; please remove it' %
                  (OQ, os.readlink(OQ)))
@@ -67,12 +81,15 @@ def install():
             cfg.write('[dbserver]\nport = %d\n' % DBSERVER_PORT)
         print('Created %s' % OQ_CFG)
 
-    # if the installation went correctly, create a symlink to oq
+    # create symlink to oq
     if not os.path.exists(OQ):
         os.symlink('%s/bin/oq' % VENV, OQ)
 
     # start the DbServer
     subprocess.check_call(['%s/bin/oq' % VENV, 'dbserver', 'start'])
+
+    # start the WebUI
+    subprocess.check_call(['%s/bin/oq' % VENV, 'webui', 'start'])
 
 
 if __name__ == '__main__':
